# files.go
package main

import (
	"errors"
	"fmt"
	"net/http"
	"os"
	"path/filepath"
	"strconv"
	"strings"

	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
	"gorm.io/gorm"
)

// Paginate https://gorm.io/docs/scopes.html#Pagination
func Paginate(r *http.Request) func(db *gorm.DB) *gorm.DB {
	return func(db *gorm.DB) *gorm.DB {
		q := r.URL.Query()
		page, _ := strconv.Atoi(q.Get("page"))
		if page <= 0 {
			page = 1
		}

		pageSize, _ := strconv.Atoi(q.Get("page_size"))
		switch {
		case pageSize > 100:
			pageSize = 100
		case pageSize <= 0:
			pageSize = 10
		}

		offset := (page - 1) * pageSize
		return db.Offset(offset).Limit(pageSize)
	}
}

// @Summary Get a list of files for the authenticated user
// @Description Get a list of files for the authenticated user. The user is identified by the token in the cookie.
// @Tags files
// @Produce  json
// @Success 200 {array} File
// @Failure 400
// @Router /files [get]
func (app *App) getFiles(c *gin.Context) {
	user := c.MustGet(ContextUserKey).(*User)

	var files []File
	result := app.db.Scopes(Paginate(c.Request)).Preload("Tags", nil).Where(&File{UserId: user.ID}).Find(&files)
	if result.Error != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": result.Error.Error()})
	}
	c.JSON(http.StatusOK, files)
	return
}

func (app *App) doesFileNameExist(ctx *gin.Context, fileName string) bool {
	_, err := gorm.G[File](app.db).Where("file_path LIKE ?", fileName).First(ctx)
	if err != nil {
		if errors.Is(err, gorm.ErrRecordNotFound) {
			return false
		}
		panic(err)
	} else {
		return true
	}
}

func (app *App) generateUniqueFileName(ctx *gin.Context) string {
	uniqueName := uuid.New().String()

	// Timestamp?

	// I'm not sure if this is necessary, but I'm not sure if the uuid is really guaranteed to be unique
	for app.doesFileNameExist(ctx, uniqueName) {
		uniqueName = uuid.New().String()
	}

	return uniqueName
}

// @Summary Create a new file
// @Description Create a new file for the authenticated user. The user is identified by the token in the cookie.
// @Tags files
// @Accept  multipart/form-data
// @Produce  json
// @Param file formData file true "File to upload"
// @Param name formData string true "Name of the file"
// @Param description formData string false "Description of the file"
// @Param tags formData string false "Comma-separated list of tag IDs"
// @Success 200 {object} File
// @Failure 400
// @Router /files [post]
func (app *App) createFile(c *gin.Context) {

	user := c.MustGet(ContextUserKey).(*User)

	uploadedFile, err := c.FormFile("file")
	fileName := c.PostForm("name")
	fileDescription := c.DefaultPostForm("description", "")

	// Parse tags
	tagsString := c.DefaultPostForm("tags", "")
	var parsedTags []Tag
	if tagsString != "" {
		tagIds := strings.Split(tagsString, `,`)

		for _, id := range tagIds {
			uintId32, err := strconv.ParseUint(id, 10, 32)
			if err != nil {
				panic(err)
			}
			parsedTags = append(parsedTags, Tag{
				ID: uint(uintId32),
			})
		}
	}

	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	uploadPath := os.Getenv("UPLOAD_PATH")
	if uploadPath == "" {
		uploadPath = "./files"
	}

	uniqueFileName := filepath.Base(app.generateUniqueFileName(c))

	err = c.SaveUploadedFile(uploadedFile, fmt.Sprintf("%s/%s", uploadPath, uniqueFileName))
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	file := File{Name: fileName, Description: fileDescription, FilePath: uniqueFileName, UserId: user.ID}
	err = gorm.G[File](app.db).Create(
		c,
		&file,
	)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	// Associate tags with the file
	if len(parsedTags) > 0 {
		// You need to use the classic GORM API for associations. ???
		err = app.db.Model(&file).Association("Tags").Append(parsedTags)
		if err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
			return
		}
	}

	// I'm querying the database here to get the updatedAt and createdAt timestamps
	var fileFromDatabase File
	tx := app.db.Model(&File{}).Preload("Tags").First(&fileFromDatabase, file.ID)
	if tx.Error != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": tx.Error})
		return
	}

	c.JSON(http.StatusOK, fileFromDatabase)
}

// @Summary Get a file by ID
// @Description Get a file by ID for the authenticated user. The user is identified by the token in the cookie.
// @Tags files
// @Produce  json
// @Param id path int true "File ID"
// @Success 200 {object} File
// @Failure 400
// @Router /files/{id} [get]
func (app *App) getFile(c *gin.Context) {
	user := c.MustGet(ContextUserKey).(*User)

	// TODO: There must be a better way
	fileId, err := strconv.Atoi(c.Param("id"))
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	fileIdAsUint := uint(fileId)

	file, err := gorm.G[File](app.db).Preload("Tags", nil).Where(&File{UserId: user.ID, ID: fileIdAsUint}).First(c)

	if errors.Is(err, gorm.ErrRecordNotFound) {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, file)
}

// @Summary Delete a file by ID
// @Description Delete a file by ID for the authenticated user. The user is identified by the token in the cookie.
// @Tags files
// @Produce  json
// @Param id path int true "File ID"
// @Success 200 {object} object
// @Failure 400
// @Router /files/{id} [delete]
func (app *App) deleteFile(c *gin.Context) {
	user := c.MustGet(ContextUserKey).(*User)

	fileId, err := strconv.Atoi(c.Param("id"))
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	fileIdAsUint := uint(fileId)

	_, err = gorm.G[File](app.db).Where(&File{UserId: user.ID, ID: fileIdAsUint}).Delete(c)

	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, gin.H{"success": true})
}

// @Summary Update a file by ID
// @Description Update a file by ID for the authenticated user. The user is identified by the token in the cookie.
// @Tags files
// @Accept  json
// @Produce  json
// @Param id path int true "File ID"
// @Param file body File true "File object that needs to be updated"
// @Success 200 {object} object
// @Failure 400
// @Router /files/{id} [patch]
func (app *App) updateFile(c *gin.Context) {
	user := c.MustGet(ContextUserKey).(*User)

	fileId, err := strconv.Atoi(c.Param("id"))
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	fileIdAsUint := uint(fileId)

	var file File
	if err := c.BindJSON(&file); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	_, err = gorm.G[File](app.db).Where(&File{UserId: user.ID, ID: fileIdAsUint}).Updates(c, file)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	fetchedFile, err := gorm.G[File](app.db).Where(&File{UserId: user.ID, ID: fileIdAsUint}).First(c)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	// Associate tags with the file
	if len(file.Tags) > 0 {
		// You need to use the classic GORM API for associations. ??? maybe not
		fmt.Println(file.Tags)
		err = app.db.Model(&fetchedFile).Association("Tags").Replace(file.Tags)
		if err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
			return
		}
	}

	c.JSON(http.StatusOK, gin.H{"success": true})
}